* release/0.2.9:
  fix database dump tasks
  improve setup script
  fix symlinking not working as expected on linux
  update deps
  update npm deps
  update dependencies

+++ app/templates/setup.sh
@@ -5,17 +5,24 @@
 if ! [ -L styleguide/public/assets ]; then
 	echo "assets not yet symlinked from styleguide - symlinking"
 	cd styleguide/public
-	ln -s ../../src/site/templates/assets
+	ln -s ../../src/site/templates/assets assets
+fi
+
+# add private.json if it doesn't yet exist
+if [ -f private-sample.json ] && ! [ -f private.json ]; then
+	echo "private.json not present - making a copy using private-sample.json"
+	cp private-sample.json private.json
 fi
 
 # initialise bower if no bower.json exists
 if ! [ -f bower.json ]; then
-	echo "bower.json not present - initialising bower"
-	bower init
-fi
-
-# add private.json if it doesn't yet exist
-if ! [ -f private.json ]; then
-	echo "private.json no present - copying private-sample.json"
-	cp private-sample.json private.json
+	if ! type bower > /dev/null; then
+		echo "you don't have bower installed"
+		echo "run npm install bower -g && npm install"
+		echo "aborting..."
+		exit 1
+	else
+		echo "bower.json not present - initialising bower"
+		bower init
+	fi
 fi
